refactor: rename user.entity to user.interface
