+++ src/modules/auth/models/AuthRequest.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
-import { User } from 'src/modules/user/user.entity';
+import { User } from 'src/modules/user/user.interface';
 
 export interface AuthRequest extends Request {
   user: User;
