--- src/modules/user/user.entity.ts
@@ -1,10 +1,0 @@
-export class User {
-  id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-  birthday: string;
-  confirmation_key: string;
-  confirmation_sit: number;
-}
